pyproject.toml
[tool.poetry]
name = "tradebot"
version = "0.1.0"
description = ""
authors = ["daniilTrukhv <daniiltrukhov256@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
binance-futures-connector = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
